+++ LocadoraVeiculos.Dominio/Modulo Plano/Plano.cs
@@ -8,13 +8,13 @@ namespace LocadoraVeiculos.Dominio.Modulo_Plano
         public GrupoVeiculo Grupo { get; set; }
         public Guid GrupoVeiculoId { get; set; } // tem que ter isso com o Orm
 
-        public float ValorDiario_Diario { get; set; }
-        public float ValorPorKm_Diario { get; set; }
+        public double ValorDiario_Diario { get; set; }
+        public double ValorPorKm_Diario { get; set; }
 
-        public float ValorDiario_Livre { get; set; }
+        public double ValorDiario_Livre { get; set; }
 
-        public float ValorDiario_Controlado { get; set; }
-        public float ValorPorKm_Controlado { get; set; }
+        public double ValorDiario_Controlado { get; set; }
+        public double ValorPorKm_Controlado { get; set; }
         public int LimiteQuilometragem_Controlado { get; set; }
 
         public string NomeGrupo
